+++ src/main/php/xp/compiler/emit/php/Emitter.class.php
@@ -16,6 +16,7 @@ use xp\compiler\types\Property;
 use xp\compiler\types\Operator;
 use xp\compiler\types\Indexer;
 use xp\compiler\types\Constant;
+use xp\compiler\types\Parameter;
 use xp\compiler\ast\ParseTree;
 use xp\compiler\ast\VariableNode;
 use xp\compiler\ast\TypeDeclarationNode;
@@ -1421,7 +1422,7 @@ abstract class Emitter extends \xp\compiler\emit\Emitter {
         }
       }
 
-      $signature[]= new TypeName($ptr->name());
+      $signature[]= new Parameter($param['name'], new TypeName($ptr->name()), isset($param['default']) ? $param['default'] : null);
       $genericParams.= ', '.$t->compoundName();
       $this->metadata[0][1][$this->method[0]][DETAIL_ARGUMENTS][$i]= $ptr->name();
       
@@ -1797,10 +1798,16 @@ abstract class Emitter extends \xp\compiler\emit\Emitter {
       $this->properties[0][$name][$property->name]= array($property->type, $statements);
     }
 
+    $type= $this->resolveType($property->type);
+    $this->metadata[0][0][$property->name]= array(
+      DETAIL_ANNOTATIONS  => array('type' => $type->name()),
+      DETAIL_PROPERTY     => $property->modifiers
+    );
+
     // Register type information
     $p= new Property();
     $p->name= $property->name;
-    $p->type= new TypeName($this->resolveType($property->type)->name());
+    $p->type= new TypeName($type->name());
     $p->modifiers= $property->modifiers;
     $this->types[0]->addProperty($p);
   }    
@@ -1862,7 +1869,8 @@ abstract class Emitter extends \xp\compiler\emit\Emitter {
       $this->leave();
     }
     
-    // Declare auto-properties as private with null as initial value
+    // Declare auto-properties as private with null as initial value. Declare a
+    // public static member with all properties' names and types as hashmap.
     foreach ($auto as $name => $none) $b->append('private $__·'.$name.'= null;');
   }