+++ src/test/php/net/xp_lang/tests/types/TypeInstanceTest.class.php
@@ -4,6 +4,7 @@ use xp\compiler\types\TypeInstance;
 use xp\compiler\types\TypeReflection;
 use xp\compiler\types\TypeName;
 use xp\compiler\types\Types;
+use xp\compiler\types\Parameter;
 use lang\XPClass;
 
 /**
@@ -13,43 +14,36 @@ use lang\XPClass;
  */
 class TypeInstanceTest extends \unittest\TestCase {
 
-  /**
-   * Test
-   *
-   */
   #[@test]
-  public function runnableInterfaceInstanceHasToStringMethod() {
+  public function runnable_interface_instance_tostring_exists() {
     $this->assertTrue(create(new TypeInstance(new TypeReflection(XPClass::forName('lang.Runnable'))))->hasMethod('toString'));
   }
 
-  /**
-   * Test
-   *
-   */
   #[@test]
-  public function runnableInterfaceInstanceToString() {
+  public function runnable_interface_instance_tostring_return() {
     $m= create(new TypeInstance(new TypeReflection(XPClass::forName('lang.Runnable'))))->getMethod('toString');
     $this->assertEquals(new TypeName('string'), $m->returns);
   }
 
-  /**
-   * Test getExtensions() method
-   *
-   */
   #[@test]
-  public function objectClassHasNoExtensionMethods() {
+  public function runnable_interface_instance_equals_parameters() {
+    $m= create(new TypeInstance(new TypeReflection(XPClass::forName('lang.Runnable'))))->getMethod('equals');
+    $this->assertEquals(
+      array(new Parameter('cmp', new TypeName('lang.Generic'))),
+      $m->parameters
+    );
+  }
+
+  #[@test]
+  public function object_class_has_no_extension_methods() {
     $this->assertEquals(
       array(), 
       create(new TypeInstance(new TypeReflection(XPClass::forName('lang.Object'))))->getExtensions()
     );
   }
 
-  /**
-   * Test getExtensions() method
-   *
-   */
   #[@test]
-  public function extensionMethod() {
+  public function extension_methods() {
     $extensions= create(new TypeInstance(new TypeReflection(XPClass::forName('net.xp_lang.tests.types.ArraySortingExtensions'))))->getExtensions();
 
     $this->assertEquals(1, sizeof($extensions));