+++ src/test/php/net/xp_lang/tests/types/TypeReflectionTest.class.php
@@ -2,7 +2,10 @@
 
 use xp\compiler\types\TypeReflection;
 use xp\compiler\types\TypeName;
+use xp\compiler\types\Parameter;
 use xp\compiler\types\Types;
+use xp\compiler\ast\IntegerNode;
+use xp\compiler\ast\NullNode;
 use lang\XPClass;
 use lang\ClassLoader;
 
@@ -143,7 +146,13 @@ class TypeReflectionTest extends \unittest\TestCase {
     $method= create(new TypeReflection(XPClass::forName('lang.types.String')))->getMethod('substring');
     $this->assertEquals(new TypeName('lang.types.String'), $method->returns);
     $this->assertEquals('substring', $method->name);
-    $this->assertEquals(array(new TypeName('int'), new TypeName('int')), $method->parameters);
+    $this->assertEquals(
+      array(
+        new Parameter('start', new TypeName('int')),
+        new Parameter('length', new TypeName('int'), new IntegerNode(0))
+      ),
+      $method->parameters
+    );
     $this->assertEquals(MODIFIER_PUBLIC, $method->modifiers);
   }
 
@@ -300,22 +309,45 @@ class TypeReflectionTest extends \unittest\TestCase {
     );
   }
 
-  /**
-   * Test "self" used in parameter type
-   *
-   */
   #[@test]
   public function selfParameterType() {
     $builder= create(new TypeReflection(XPClass::forName('net.xp_lang.tests.types.Builder')));
     $this->assertEquals(
-      new TypeName('net.xp_lang.tests.types.Builder'),
+      new Parameter('self', new TypeName('net.xp_lang.tests.types.Builder'), new NullNode()),
       $builder->getMethod('create')->parameters[0]
     );
   }
 
-  /**
-   * Test enum member type
-   */
+  #[@test]
+  public function parameter_with_array_default() {
+    $cl= ClassLoader::defineClass('TypeReflectionTest_'.$this->name, 'lang.Object', array(), '{
+      /** @param string[] param */
+      public function fixture($param= array()) { }
+    }');
+    $this->assertEquals(
+      new Parameter('param', new TypeName('string[]'), new \xp\compiler\ast\ArrayNode(array(
+        'type'   => new TypeName('string[]'),
+        'values' => array()
+      ))),
+      create(new TypeReflection($cl))->getMethod('fixture')->parameters[0]
+    );
+  }
+
+  #[@test]
+  public function parameter_with_map_default() {
+    $cl= ClassLoader::defineClass('TypeReflectionTest_'.$this->name, 'lang.Object', array(), '{
+      /** @param [:string] param */
+      public function fixture($param= array()) { }
+    }');
+    $this->assertEquals(
+      new Parameter('param', new TypeName('[:string]'), new \xp\compiler\ast\MapNode(array(
+        'type'     => new TypeName('[:string]'),
+        'elements' => array()
+      ))),
+      create(new TypeReflection($cl))->getMethod('fixture')->parameters[0]
+    );
+  }
+
   #[@test]
   public function enumMemberType() {
     $cl= ClassLoader::defineClass('TypeReflectionTest_Enum', 'lang.Enum', array(), '{
@@ -327,9 +359,6 @@ class TypeReflectionTest extends \unittest\TestCase {
     );
   }
 
-  /**
-   * Test enum member type
-   */
   #[@test]
   public function abstractEnumMemberType() {
     $cl= ClassLoader::defineClass('TypeReflectionTest_AbstractEnum', 'lang.Enum', array(), '{