+++ ChangeLog.md
@@ -3,6 +3,19 @@ XP Compiler ChangeLog
 
 ## ?.?.? / ????-??-??
 
+## 1.11.0 / 2013-10-05
+
+* Heads up: Changed xp.compiler.types.Types implementations to return an 
+  array of Parameter objects holding name, type and optional default value 
+  instead of just the parameters' types - (@thekid)
+* Made error message more verbose when syntax cannot be determined from
+  file referenced by command line argument - (@thekid)
+* Fixed Fatal error: Class 'xp\compiler\ClassLoader' not found when using
+  command line argument `-cp` - (@thekid)
+
+## 1.10.4 / 2013-06-02
+
+* Fixed properties inconsistency - see pull request #35 (@thekid)
 * Implemented RFC 250: Static member import syntax change (@thekid)
 
 ## 1.10.3 / 2013-05-27